Refactored PlanDetailsTable.tsx - separated logic to a new file - usePlanDetailsTableLogic.ts, in PlanDetailsTable.tsx left only the view.
